Provide a listener_add limitation work around.
This involves keeping a shadow copy of the Vim buffer's contents and
modifying it as Vim reports changes. Changes are not applied to the
Tree-sitter tree until a SafeAgainState event occurs. My home/tenous
theory is that the SafeAgainState event only triggers when shadow has
'caught up' with the Vim buffer.

A comparison between shadow and actual buffer is made. If that fails
then the problem is logged and a full parse triggered. This part of the
mechanism will only be tested after I have run with in inplace for some
time, assuming the tracking is not solid.